2025-Jun-15, casting mid * mid to long, catching overflow case

+++ Java/Day3_2_Sqrt.java
@@ -12,7 +12,7 @@ public class Day3_2_Sqrt {
 			mid = (low + high) / 2;
 			if (mid * mid == num)
 				return (mid);
-			else if (mid * mid <= num)
+			else if ((long) mid * mid <= num)
 				low = mid + 1;
 			else
 				high = mid - 1;
